fix: removes indirect otel dependencies
Many Temporal applications depend on go.temporal.io/server which is
pinned to a fairly old otel dependency. This can cause conflicts with
the pkg/expression package required by this plugin, which uses the
Bloblang subcomponent of the github.com/benthosdev/benthos/v4 project
that has a tight coupling with the a newer version of the otel
dependency. As the only dependency of the plugin is for the pure subset
of the Bloblang functionality, this change replaces the benthos
dependency with the github.com/cludden/benthos/v4@v4.24.0-no-otel.1 fork
that removes everything but the necessary Bloblang subset.