fix: fixes activity option task queue behavior
The previous implementation initiaized the ActivityOptions for the
ExecuteActivity method using the built-in GetActivityOptions helper.
Unfortunately, this helper always returns an ActivityOptions value with
a non-empty task queue that defaults to the workflow task queue, which
prevents the task queue defaulting logic from executing. This change
instead initializes an empty ActivityOptions value and sets the default
task queue to either the value specified by the proto task_field option
or the workflow task queue if no proto value is defined.

closes: #37