Improve strcase behavior, expand cli options
This builds on top of #89 by replacing the existing usage of the
`iancoleman/strcase` third-party library with a local strcase package
that better supports the new ignore-acronyms plugin option and also
changes the flags behavior to use semicolon delimiters for better
ergonomics.

In order to prevent backwards incompatible changes to generated code,
the protobuf options have been expanded to support custom overrides to
various cli command and field names, aliases, and usage.